requirements.txt
Django>=4.2,<5.0
requests>=2.31.0
beautifulsoup4>=4.12.2
django-apscheduler>=0.6.2
apscheduler>=3.10.1
python-dotenv>=1.0.0
djangorestframework>=3.16.0
drf-spectacular>=0.26.4
psycopg2-binary>=2.9.9